requirements.txt
flet==0.25.2
sqlalchemy==2.0.36
pymysql==1.1.1
requests==2.32.3
python-dotenv==1.0.1
qrcode==7.4.2
pillow==10.4.0
reportlab==4.2.5
pyinstaller==6.11.1
